requirements.txt
streamlit
pandas
matplotlib
numpy
qrcode
pillow